+++ rust/dbn/src/decode/dbn/async.rs
@@ -656,14 +656,15 @@ mod tests {
             dbn::{AsyncEncoder, AsyncRecordEncoder},
             DbnEncodable,
         },
-        rtype, CbboMsg, Error, ErrorMsg, ImbalanceMsg, InstrumentDefMsg, MboMsg, Mbp10Msg, Mbp1Msg,
-        OhlcvMsg, RecordHeader, Result, Schema, StatMsg, StatusMsg, TbboMsg, TradeMsg, WithTsOut,
+        rtype, Cmbp1Msg, Error, ErrorMsg, ImbalanceMsg, InstrumentDefMsg, MboMsg, Mbp10Msg,
+        Mbp1Msg, OhlcvMsg, RecordHeader, Result, Schema, StatMsg, StatusMsg, TbboMsg, TradeMsg,
+        WithTsOut,
     };
 
     #[rstest]
     #[case::mbo(Schema::Mbo, MboMsg::default())]
     #[case::trades(Schema::Trades, TradeMsg::default())]
-    #[case::tbbo(Schema::Cbbo, CbboMsg::default_for_schema(Schema::Cbbo))]
+    #[case::tbbo(Schema::Cmbp1, Cmbp1Msg::default_for_schema(Schema::Cmbp1))]
     #[case::tbbo(Schema::Tbbo, TbboMsg::default())]
     #[case::mbp1(Schema::Mbp1, Mbp1Msg::default())]
     #[case::mbp10(Schema::Mbp10, Mbp10Msg::default())]
@@ -708,7 +709,7 @@ mod tests {
     #[rstest]
     #[case::mbo(Schema::Mbo, MboMsg::default())]
     #[case::trades(Schema::Trades, TradeMsg::default())]
-    #[case::cbbo(Schema::Cbbo, CbboMsg::default_for_schema(Schema::Cbbo))]
+    #[case::cbbo(Schema::Cmbp1, Cmbp1Msg::default_for_schema(Schema::Cmbp1))]
     #[case::tbbo(Schema::Tbbo, TbboMsg::default())]
     #[case::mbp1(Schema::Mbp1, Mbp1Msg::default())]
     #[case::mbp10(Schema::Mbp10, Mbp10Msg::default())]