+++ rust/dbn/src/record_ref.rs
@@ -166,8 +166,9 @@ impl<'a> From<&'a RecordEnum> for RecordRef<'a> {
             RecordEnum::Error(rec) => Self::from(rec),
             RecordEnum::SymbolMapping(rec) => Self::from(rec),
             RecordEnum::System(rec) => Self::from(rec),
-            RecordEnum::Cbbo(rec) => Self::from(rec),
+            RecordEnum::Cmbp1(rec) => Self::from(rec),
             RecordEnum::Bbo(rec) => Self::from(rec),
+            RecordEnum::Cbbo(rec) => Self::from(rec),
         }
     }
 }
@@ -187,8 +188,9 @@ impl<'a> From<RecordRefEnum<'a>> for RecordRef<'a> {
             RecordRefEnum::Error(rec) => Self::from(rec),
             RecordRefEnum::SymbolMapping(rec) => Self::from(rec),
             RecordRefEnum::System(rec) => Self::from(rec),
-            RecordRefEnum::Cbbo(rec) => Self::from(rec),
+            RecordRefEnum::Cmbp1(rec) => Self::from(rec),
             RecordRefEnum::Bbo(rec) => Self::from(rec),
+            RecordRefEnum::Cbbo(rec) => Self::from(rec),
         }
     }
 }